+++ config/hypr/scripts/KeyHints.sh
@@ -57,7 +57,7 @@ GDK_BACKEND=$BACKEND yad \
 "ALT Print" "Screenshot active window" "active window only" \
 "CTRL ALT P" "power-menu" "(wlogout)" \
 "CTRL ALT L" "screen lock" "(swaylock screensaver, hyprlock fallback)" \
-" SHIFT L" "screensaver hack picker" "(lock background, Alt+P previews)" \
+" SHIFT L" "screensaver hack picker" "(Alt+P preview, Alt+C edit hack flags)" \
 "CTRL ALT Del" "Hyprland Exit" "(NOTE: Hyprland Will exit immediately)" \
 " SHIFT F" "Fullscreen" "Toggles to full screen" \
 " CTL F" "Fake Fullscreen" "Toggles to fake full screen" \